extract interface from AssetsAndCashService
